+++ src/SamplesApp/Benchmarks.Shared/Suite/Windows_UI/ColorBench/XamlControlsResourcesCreationBenchmark.cs
@@ -1,5 +1,4 @@
-﻿#pragma warning disable CS0219 // Variable is assigned but its value is never used
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Text;
 using BenchmarkDotNet.Attributes;
@@ -44,4 +43,3 @@ public class ColorBenchmark
 		var b = _color1.Equals(_color2);
 	}
 }
-#pragma warning restore CS0219 // Variable is assigned but its value is never used
