+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.cs
@@ -42,7 +42,7 @@ namespace Windows.UI.Xaml.Controls
 		private readonly BorderLayerRenderer _borderRenderer;
 
 		public Border()
-        {
+		{
 			_borderRenderer = new BorderLayerRenderer(this);
 		}
